+++ components/ui/team-item.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { TouchableOpacity } from 'react-native-gesture-handler'
 import { List } from 'react-native-paper'
 
-import config from '../config'
-import i18n from '../lib/i18n'
-import { formatDates } from '../lib/moment'
+import config from '../../config'
+import i18n from '../../lib/i18n'
+import { formatDates } from '../../lib/moment'
 
 const TeamItem = ({ item, onPress }) => {
   const { team, event } = item
